backend/requirements.txt
django==4.0.4
djangorestframework==3.13.1
django-cors-headers==3.12.0
Pillow==9.1.1